// client/pageComponents/CreateSocialForm/createSocialForm.schema.ts
export interface SocialData {
  title?: string;
  venue?: string;
  capacity?: string;
  price?: string;
  description?: string;
  isManualApprove?: boolean;
  privacy?: string;
  tags?: string[];
  banner?: string;
  date?: any;
  time?: any;
  startAt?: any;
  dateFormat?: any;
  timeFormat?: any;
}